refactor: fix name of `getOptionalLinkProperties` file
