+++ src/lib/getOptionalLinkProperties.ts
